requirements.txt
numpy>=1.26.0
open3d>=0.19.0
Pillow>=10.0.0
PyYAML>=6.0
transformers>=4.45.0
torchvision>=0.19.0
